+++ Sample.Domain/QA/Questions/Question.cs
@@ -5,7 +5,7 @@ using Sample.Domain.Users;
 
 namespace Sample.Domain.QA.Questions;
 
-[EntityType((int)SampleEntityTypes.Question)]
+[EntityType((byte)SampleEntityTypes.Question)]
 public class Question : AggregateRoot
 {
     private Question()
